fix: use export type * to avoid empty exports file, closes #2
It seems the new astro/vite version has trouble importing and
re-exporting empty exports files.
The types.ts contains only type definitions and compiles down to
just "exports {}".

This breaks module loading by astro/vite.

Updating to typescript 5 required to support the new "export type *"
syntax which avoids to emit the re-export, that is, the "export *"
for an empty file.